+++ src/notification.h
@@ -76,7 +76,7 @@ class notification_box_t : public gui_object_t {
   void draw_background(int width, int height, int sprite);
   void draw_string(int x, int y, const std::string &str);
   void draw_map_object(int x, int y, int sprite);
-  int get_player_face_sprite(int face);
+  unsigned int get_player_face_sprite(size_t face);
   void draw_player_face(int x, int y, int player);
 
   virtual void internal_draw();
